src/test-app/requirements.txt
anyio==3.5.0
appdirs==1.4.4
appnope==0.1.2
asgiref==3.5.0
astroid==2.9.3
asttokens==2.0.5
attrs==20.3.0
backcall==0.2.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
coverage==5.5
debugpy==1.5.1
decorator==5.0.7
Deprecated==1.2.13
distlib==0.3.1
Django==4.0.2
djangorestframework==3.13.1
entrypoints==0.4
executing==0.8.3
fastapi==0.75.0
filelock==3.0.12
Flask==1.1.2
h11==0.13.0
idna==2.10
iniconfig==1.1.1
ipykernel==6.9.2
ipython==8.1.1
ipython-genutils==0.2.0
isort==5.10.1
itsdangerous==1.1.0
jaeger-client==4.4.0
jedi==0.18.0
Jinja2==2.11.3
jupyter-client==7.1.2
jupyter-core==4.9.2
lazy-object-proxy==1.7.1
MarkupSafe==1.1.1
matplotlib-inline==0.1.3
mccabe==0.6.1
nest-asyncio==1.5.4
numpy==1.22.3
opensky-sv==0.0.1
opentelemetry-api==1.11.0
opentelemetry-exporter-jaeger-thrift==1.11.0
opentelemetry-instrumentation==0.30b0
opentelemetry-instrumentation-asgi==0.30b0
opentelemetry-instrumentation-fastapi==0.30b0
opentelemetry-instrumentation-requests==0.30b0
opentelemetry-sdk==1.11.0
opentelemetry-semantic-conventions==0.30b0
opentelemetry-util-http==0.30b0
opentracing==2.4.0
packaging==20.9
pandas==1.4.2
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.1
pluggy==0.13.1
prompt-toolkit==3.0.18
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.10.0
pydantic==1.9.0
Pygments==2.8.1
pylint==2.12.2
pyparsing==2.4.7
pytest==6.2.3
python-dateutil==2.8.2
pytz==2021.3
pyzmq==22.3.0
requests==2.25.1
six==1.15.0
sniffio==1.2.0
sqlparse==0.4.2
stack-data==0.2.0
starlette==0.17.1
threadloop==1.0.2
thrift==0.13.0
toml==0.10.2
tornado==6.1
traitlets==5.1.1
typing_extensions==4.1.1
urllib3==1.26.4
uvicorn==0.17.6
virtualenv==20.4.3
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.13.3
